requirements.txt
# requirements.txt

gradio
langchain
langchain-openai
langchain-upstage
langchain-ollama
langchain-community
faiss-cpu
python-dotenv
numpy
scikit-learn
ollama